+++ dashboard/membre.php
@@ -24,7 +24,7 @@ $membre = $requete->fetchAll(PDO::FETCH_ASSOC); ?>
     <div class="page-sub-header">
     <h3 class="page-title">membres</h3>
     <ul class="breadcrumb">
-    <li class="breadcrumb-item"><a href="students.html">membre</a></li>
+    <li class="breadcrumb-item"><a href="dashboard.php">dashboard</a></li>
     <li class="breadcrumb-item active">Tous les membres</li>
     </ul>
     </div>
@@ -89,7 +89,7 @@ $membre = $requete->fetchAll(PDO::FETCH_ASSOC); ?>
     <td class="desc"><?= $mbr['descriptions'] ?></td>
     <td class="text-end">
     <div class="actions ">
-    <a href="delete-membre.php?id=<?php echo $mbr['id'];?>" class="btn btn-sm bg-success-light me-2 ">
+    <a href="delete-membre.php?id=<?php echo $mbr['id'];?>" class="btn btn-sm bg-success-light me-2 "onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet element ?')">
     <i class="feather-delete"></i>
     </a>
     <a href="edit-membre.php?id=<?php echo $mbr['id'];?>" class="btn btn-sm bg-danger-light">
